feat: delivery option filter UI sfs 2549 (#2930)
## What's the purpose of this pull request?

When Delivery Promise enabled:

- Adds Delivery Options Filters component in PLP
- List Delivery Options available to filter
- Enable add custom label for Delivery Option section via hCMS

note: Button actions and facets filters is not fully integrated yet.

## How it works?

List Delivery Options available to filter according to the postalCode
provided (mocked for now)
| Desktop | Mobile |

|-------------------------------------------------------------------------------------------|-------------------------------------------------------------------------------------------|
|
![image](https://github.com/user-attachments/assets/2ec47c90-42e5-4411-b334-d70b9d4138bc)
|
![image](https://github.com/user-attachments/assets/06b9eed3-f982-4239-8c70-da4d0bb65f54)
|


## How to test it?

DO isn't in production in the IS API yet, so we have to test this PR
using the mock IS created.
To use the mock, make sure that the workspace vtexis (vtex login vendemo
-w vtexis) has the app vtex.intelligent-search-api installed in the
version 0.51.0-mock.1 (vtex ls). Then, make some changes in the
`packages/core/discovery.config.default.js` and
`packages/api/src/platforms/vtex/clients/search/index.ts:`

Use the vendemo account, `BRA` country and `pt-BR` locale. You can test
the behavior with DP enabled and disabled.
Add the workspace to the base URL in index.ts, it should be:

```
const base = `https://vtexis--${account}.${environment}.com.br/api/io`
```

1. Navigate to a PLP, you should be able to see a Delivery section in
the Filters component. (Also try in the mobile resolution)
2. If no postalCode provided, you should be only see `Set Location`
button
3. If postalCode provided, you should be able to see the list of
available Delivery Options for that postalCode (mocked for now)

### Starters Deploy Preview

I'd have to commit the change in the API so the preview would work,
otherwise it won't be using the workspace. To avoid committing it, I
suggest that we test only locally using the core.