feat: Allow unavailable items on cart (#2938)
## What's the purpose of this pull request?

The intention of this PR is to add the `enableUnavailableItemsOnCart`
feature flag to enable merchants to handle scenarios where items become
unavailable for some reason (e.g. after changing pickup point) on the
client-side.

## How it works?

It adds the feature flag to be read from `discovery.config` API flags,
but with default value set to `false`.

## How to test it?

- Check if all `@faststore/api` tests are passing;
- Perform cart actions (add/remove items) and it should behave as
before, nothing should change with `enableUnavailableItemsOnCart:
false`;
- The `enableUnavailableItemsOnCart: true` scenario will not work for
FastStore for now because our default behavior is to let the backend
handle unavailable itens on cart.

### Starters Deploy Preview

vtex-sites/faststoreqa.store#837