feat: Global filter by pickup point (SFS-2458) (#2887)
Revision order:

1) #2889
2) #2890

These changes will introduce global filtering by pickup point for the
selected zip code.

We are changing the `RegionSlider` opening: the component it's a global
section now, so it should be able to be rendered everywhere in the
store, still using the `UI Provider` actions but no longer need to
import it.

Furthermore, we made adjustments to the the `RegionBar` component,
adding new data-atributes and matching new design prototype, but we keep
backward compatibility by leaving the current props and data-attributes
unaltered.

Although, the core of this implementation it's to support global
filtering (including shelves' queries) by available pickup points, and
for that we created the Delivery Promise's reducer and provider to
handle this global context through shopper navigation, leveraging from
the `useDeliveryPromise` hook towards managing all state and actions.

First, you can use the Starter preview below, but for local development
you'll need to use the `vendemo` account we already set up for Delivery
Promise: on `discovery.config` you should change the following
attributes:
- `api: { storeId: 'vendemo', ... }`;
- `session: { currency: { code: 'BRL', symbol: 'R$' }, locale: 'pt-BR',
country: 'BRA', ... }`;
- `deliveryPromise: { enabled: true, mandatory: false }`;

After that, specify one of the following zip codes (using the
`RegionButton` on the navbar) to test Delivery Promise features:
`50030-260`, `20070-001` or `04538-132`.

Then, try to set a global pickup point and see if the store name is
added on the navbar, navigate to PLP and you should see the `Delivery`
facet section, you can filter by other methods. Everything should keep
working as before. Also, try the mobile version.

vtex-sites/faststoreqa.store#834

---------

Co-authored-by: Larícia Mota <laricia.mota@vtex.com.br>
Co-authored-by: Fanny Chien <fanny.chien@vtex.com>