fix: use allowCancellation to validate cancel order button (#2942)
## What's the purpose of this pull request?

Replicate the same behavior for cancel order as in my-orders.

## How it works?

Use allowCancellation flag from userOrders query to check when to show
the cancel order button.

Based in my-orders, you can check there
[here](https://github.com/vtex/my-orders/blob/ff400736a4b6ea8eb70298c2de26a6e974489fc3/react/pages/CancelOrder.jsx#L116)
and
[here](https://github.com/vtex/my-orders/blob/ff400736a4b6ea8eb70298c2de26a6e974489fc3/react/components/Order/OrderActions.jsx#L41)

## How to test it?

Open a order and click cancel order button

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->

[JIRA](https://vtex-dev.atlassian.net/browse/SFS-2663)