fix: format date manually on MyAccountSelectedTags to work with user timezone (#2946)
## What's the purpose of this pull request?

Adjust the filtered date shown on my account list orders to be the same
as in the drawer.

## How it works?

When calling new Date() with a date in "YYYY-MM-DD" format it applies
the timezone to it.
So I am now formatting manually the date on MyAccountSelectedTags
extracting year, month and day from the string.

## How to test it?

Filter the order list with a date range

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->

[JIRA](https://vtex-dev.atlassian.net/browse/SFS-2641)

<img width="1371" height="769" alt="image"
src="https://github.com/user-attachments/assets/48c53d09-653b-4554-8650-0348a0712f66"
/>