chore: add Pickup Point's address state back (#2962)
## What's the purpose of this pull request?

When we changed the API (from Checkout to Logistics) we removed the info
related to the address state of the pickup points, because Logistics
wasn't returning this info. But in the latest deploy they added this
info so we can add it back!

## How it works?

Just adding state back 

## How to test it?

Check that the state appears in the Pickup Point Card in the list:
<img width="406" height="424" alt="Screenshot 2025-07-25 at 11 28 43"
src="https://github.com/user-attachments/assets/8c810e30-d368-4b7b-a368-8bffaf3d1d93"
/>

### Starters Deploy Preview

PR:
Preview:

## References

[Jira task](https://vtex-dev.atlassian.net/browse/SFS-2678)