feat: Better monorepo support (#2978)
## What's the purpose of this pull request?

In order to support the monorepo architecture on Homebrew, I decided to
simplify the way we establish where to put the `. faststore` directory.
Now the `.faststore` directory is created on `process.cwd()`, instead of
where the discovery.config.js is located.

For stores that are not using the monorepo, the behavior should be the
same.
For stores that **are** using the monorepo, the .faststore directory
will now be created on the monorepo root. This will also simplify the
management of the file-traced files.

## How to test it?

Install the preview version of the CLI and run `yarn build` and see if
everything works.

I've deployed it on homebrew and on Vercel, both monorepo and
standalone. Also run `yarn dev`, `yarn cms-sync` and the usual commands.

### Starters Deploy Preview

| | Single  | Monorepo |
|---|---|--|
| Vercel |
[link](https://guila-storeframework-cmbzjdjia000dur5533n76559-mlgb3zpvf.b.vtex.app/)
| [link](https://guila-1750105079139-storeframework.b.vtex.app/) |
| Homebrew |
[link](https://sfj-200006--guila-teste.vtex.app?an=storeframework) |
[link](https://sfj-000006--guila-teste.vtex.app/?an=storeframework) |

## References

https://github.com/vtex/dk-cicd-hub/pull/1063

## Checklist

<em>You may erase this after checking them all :wink:</em>

**PR Title and Commit Messages**

- [X] PR title and commit messages follow the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) specification
- Available prefixes: `feat`, `fix`, `chore`, `docs`, `style`,
`refactor`, `ci` and `test`

**PR Description**

- [X] Added a label according to the PR goal - `breaking change`, `bug`,
`contributing`, `performance`, `documentation`..

**Dependencies**

- [X] Committed the `pnpm-lock.yaml` file when there were changes to the
packages

**Documentation**

- [X] PR description
- [X] For documentation changes, ping `@Mariana-Caetano` to review and
update (Or submit a doc request)