fix: Update search page when changing terms (#2992)
## What's the purpose of this pull request?

Currently, after performing a new search, when the shopper is already in
the search page, the results are not changing. Also, the results are not
being updated when the zip code is changed.

With these changes, the search page will consider the correct term when
performing the search and will refetch the query when the zip code is
changed.

## How to test it?

- Search for `apple`. You should see `apple` results;
- Still in the search page, search again for `unbranded` and the results
should be different (showing results for `unbranded`);

### Starters Deploy Preview

dp-faststore-org/vendemo-dp#74
vtex-sites/faststoreqa.store#857

### References

- https://swr.vercel.app/docs/mutation#revalidation