fix: search update (#3015)
## What's the purpose of this pull request?

The search context wasn't being updated properly.
(And the `ProductTiles` wasn't considering the global pickup point - I
took the opportunity to fix it also)

## How it works?

I've reverted the previous fix, which was fixing a part of the issue, so
we could find a solution to fix the whole issue. And I found that since
the Search Provider had a useEffect with an empty dependency array, it
was causing the search to not be updated properly.

## How to test it?

You can test searching and navigating through PLPs. You can select
facets, and make a new search for example.

An example of issue (account `vendemo`):
1. Search for `women's microdot pink`
2. Select size L in the facet `Size`
3. Search for `women's microdot night`
4. Before the fix: the size facet selected in the previous search is
still selected
5. Before the fix: the URL isn't updated with the fuzzy values
6. Select size M in the facet `Clothes Size`
7. Before the fix: the search term in the URL search params go back to
the previous search ("pink")
8. Before the fix: if you refresh the page the search results go back to
the previous search ("pink")

After the fix the selected facets should be reseted when making a new
search, the URL should contain fuzzy values and should not go back to
previous search params.

### Starters Deploy Preview
https://github.com/dp-faststore-org/vendemo-dp/pull/84 /
https://github.com/vtex-sites/faststoreqa.store/pull/863

- Before the fix: 
- (`vendemo`) -
https://vendemo-cm9sir9v900u7z6llkl62l70j-c0yp9zvaf.b.vtex.app/
- (`storeframework`) -
https://storeframework-cm652ufll028lmgv665a6xv0g-ht2gb8snr.b.vtex.app/

- After the fix:
- (`vendemo`) -
https://vendemo-cm9sir9v900u7z6llkl62l70j-tcm1nzpoi.b.vtex.app/
- (`storeframework`) -
https://storeframework-cm652ufll028lmgv665a6xv0g-mto9dm1vc.b.vtex.app/

## References

- [Jira task](https://vtex-dev.atlassian.net/browse/SFS-2829)