chore: Update Org Account URL (#3045)
## What's the purpose of this pull request?

The url for the Org Account is being updated in the following PR:
https://github.com/vtex/faststore-plugin-buyer-portal/pull/154

This update makes sure that the redirects inside FastStore stay
consistent

Both PRs should be deployed in sync.

## How it works?

<!--- Tell us the role of the new feature, or component, in its context.
Provide details about what you have implemented and screenshots if
applicable. --->

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->