fix: signin button flickering - SFS-2862 (#3078)
## What's the purpose of this pull request?

This PR aims to prevent the flickering behavior from the SignIn Button.

| Before | After |
|--------|--------|
| ![beforeOct-22-2025
20-33-56](https://github.com/user-attachments/assets/00a76b97-9805-49c0-8bfb-2c0abfba0941)
|![afterOct-22-2025
20-43-25](https://github.com/user-attachments/assets/fb5098e6-fb67-4e29-b533-2f0e2484ee8b)|

## How it works?

It reduces the time from the debounced handler to retrieve persisted
indexed DB values to 100ms. Preventing more race conditions and also
improving loading time (performance).

It also creates a new hook to control the validateSession stability; the
`isValidation` is not trustworthy.


## How to test it?

As a logged-in user, you can navigate through different pages and check
the sign-in button behavior.

### Starters Deploy Preview

TBD

---------

Co-authored-by: Lucas Feijó <lucas.portela@vtex.com>