feat: Add Breadcrumb fallback label to hCMS (#3099)
## What's the purpose of this pull request?

Add the fallback label to hCMS so it can be editable. We need it there
so it can be translated through the CMS.

⚠️ Breaking change: there is no more a hardcoded fallback label, the
message necessarily has to come through CMS.
This PR will be merged in the `feat/multilanguage` feature branch, which
will be available only in the next faststore major version.

## How it works?

It was hardcoded, so adding it to the hCMS schema allows it to be
editable.

## How to test it?

In a Search page, observe the breadcrumb, it should have the value
configured through the hCMS.
<img width="1028" height="501" alt="Screenshot 2025-11-04 at 10 27 49"
src="https://github.com/user-attachments/assets/30d38c18-202c-4fa7-aeb5-5e8416019b4f"
/>

If the field is not being displayed in the hCMS, remember to cms sync
and publish the page.

<img width="1497" height="757" alt="Screenshot 2025-11-04 at 10 37 38"
src="https://github.com/user-attachments/assets/d253597b-5953-4ebd-9931-2c3afb324994"
/>

### Starters Deploy Preview

-
https://storeframework-cm652ufll028lmgv665a6xv0g-77wksgb4k.b.vtex.app/s?q=pizza
([PR](https://github.com/vtex-sites/faststoreqa.store/pull/886))

## References

- [Jira task](https://vtex-dev.atlassian.net/browse/SFS-2921)