feat: adding cache-graphql command to faststore/cli (#3122)
- faststore/cli command to calculate graphql chaced queries


You can test it by running the cli on the starter or on the core
mannualy.

Ex:
- Running on core
```
  $ pnpm i
  $ cd ./packages/core
  $ (../cli & pnpm build)
  $ node ../cli/bin/run generate-types
  $ node ../cli/bin/run cache-graphql
```

Builds the cli, and regenerated the types to create the
persisted-documents.json file and run the cache-graphql.

After running it you should see the discovery.config.default.js changed
to include all the queries names that would be cached.

---------

Co-authored-by: Fanny Chien <fanny.chien@vtex.com>