fix: adds cache busting to bypass cache when logged in/out. (#3152)
## What's the purpose of this pull request?

This PR addresses the issue where a user is logged in and still sees the
old price from when they were logged out. The same when logged in ->
out.

## How it works?

- Introduces improved cache control mechanisms for requests: browser
requests and API/graphql using session storage control.
- Cache headers are set to `private` to prevent sensitive data from
being cached publicly when the user is logged in.
- The logic for handling cookies and cache control is unified between
the client and server.


**Cache Control Improvements:**
The changes ensure that when a VTEX authentication cookie is present,
cache busting is applied, adding a new param `v=<timestamp>.`

## How to test it?

- See our [Slack channel
thread](https://vtex.slack.com/archives/C03L3CRCDC4/p1765905481238529).